+++ RELEASE_NOTES.md
@@ -4,7 +4,7 @@
 
 ### New features
 
-- Possibility to add N subpaths
+- Possibility to add N subpaths on the main form or on the table of services
 - Proper responsive on the elements
 - Fullscreen forms on mobile
 - Fullscrean table and filters